Move user entity class to entities folder
