+++ src/server/entities/user.entity.ts
